feat: extract completion_resolve to features/completion.rs
- Add resolve_completion_item<I: IndexRead>() to features/completion.rs
- completion_resolve in mod.rs becomes a 3-line adapter
- mod.rs: 731 → 696 lines
